requirements.txt
customtkinter>=5.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.8.0 